requirements.txt
pandas
openpyxl
plotly.express
streamlit

